Added separate class FileUtils
