+++ pravega-sensor-collector/src/main/java/io/pravega/sensor/collector/parquet/ParquetFileState.java
@@ -27,7 +27,7 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.pravega.sensor.collector.parquet.FileNameWithOffset;
+import io.pravega.sensor.collector.util.FileNameWithOffset;
 import io.pravega.sensor.collector.util.AutoRollback;
 import io.pravega.sensor.collector.util.TransactionCoordinator;
