+++ pravega-sensor-collector/src/main/java/io/pravega/sensor/collector/rawfile/RawFileState.java
@@ -28,6 +28,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.pravega.sensor.collector.util.AutoRollback;
+import io.pravega.sensor.collector.util.FileNameWithOffset;
 import io.pravega.sensor.collector.util.TransactionCoordinator;
 
 import static java.sql.Connection.TRANSACTION_SERIALIZABLE;
